Edit 페이지 완성

+++ day18/src/components/DiaryItem.jsx
@@ -10,7 +10,9 @@ function DiaryItem({ id, emotionId, createdDate, content }) {
   const { onDelete } = useContext(DiaryDispatchContext);
 
   const onClickDelete = () => {
-    onDelete(id);
+    if (confirm("정말 삭제하시겠습니까?")) {
+      onDelete(id);
+    } else return;
   };
 
   return (
